my_model_api/requirements.txt.txt
flask
torch>=1.8
torchvision>=0.9
Pillow
numpy